requirements.txt
fastapi
uvicorn
keras
numpy
pillow
python-multipart
h5py